Course summary image/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.25

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1

# PDF Parsing
pymupdf==1.23.5

# Image Generation
pillow==10.2.0

# Async Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
